# feat(auth): API Authorization Cache