… auth/authorizer_test.go
… common/cache_test.go
… go.sum